Move CSW specific team access methods out of iam client into ProductTeam model
